feat: add ent group

+++ database/migrations/20230113212057_group.sql
@@ -3,7 +3,7 @@
 CREATE TABLE IF NOT EXISTS group(
     id         BIGSERIAL PRIMARY KEY NOT NULL,
     name       varchar(32) NOT NULL DEFAULT '未命名',
-    owner      smallint NOT NULL DEFAULT 0,
+    owner      bigint NOT NULL DEFAULT 0,
     notice     varchar(255) NOT NULL DEFAULT '',
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP(0),
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP(0),
